+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Current (2025-10-16)
 --------------------
 
+* Added new command line tool to convert data file formats
 * Added a new experimental option for processing a record batch into root without going through a pandas data frame